+++ SimpleTableManager/SimpleTableManager/Models/IShape2Sized.cs
@@ -2,5 +2,5 @@ namespace SimpleTableManager.Models;
 
 public interface IShape2Sized : IShape1Sized
 {
-	decimal Size2 { get; set; }
+	double Size2 { get; set; }
 }
